==> 06_lazypredict_model_selection.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from lazypredict.Supervised import LazyClassifier

DATAFILE = "normalised_with_binary_cai.csv"
TARGET = "Cod_Azienda", "Anno"
df = pd.read_csv(DATAFILE).dropna(subset=[TARGET])
X = df.drop(columns=[TARGET,"Cod_Azienda","Anno"], errors="ignore"); y = df[TARGET]
X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, test_size=0.3, random_state=42)
clf = LazyClassifier(verbose=0, ignore_warnings=True, predictions=True)
models, preds = clf.fit(X_train, X_test, y_train, y_test)
models.to_csv("outputs/tables/lazypredict_benchmark.csv", index=False)
models.head()
